+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.0.2
+
+* Adjust module's namespaces
+* Correct bug in Task.__init__
+
 ## 0.0.1
 
 * Initial version